Add Apache CXF config + OGF ConnectionService WSDLs.
We will be using Apache CXF for generating Java stub code from the
OGF Connection Service WSDLs. Simply run:

    mvn generate-sources

for the stub code to be generated.

The parent package of the WSDL/SOAP stub code will be package
nl.surf.polynsi.soap. This is to differentiate it from gRPC/Protobufs
parts that will end up in nl.surf.polynsi.grpc