wsdl2java: enable processing of implicit SOAP headers.
The NSI WSDLs have an implicit SOAP header (SOAP headers defined in the
wsdl:binding but not wsdl:portType section) defined. Adding the `-exsh
true` switch to the `wsdl2java`'s configuration ensures this header is
made available in the generated code. The header is quite essential as
the whole async mechanism of NSI depends on it by virtue of the
`replyTo` attribute that's set in that custom header.

In addition we mark the generated code with the `@Generated` annotation
(`--mark-generate` switch of `wsdl2java`). This will affect stub
implementations as well. Once the stub implementations have been
implemented with real functionality one should remove the `@Generated`
annotation for those. They should remain however on all other code
(interfaces, type defs, etc)