Document building gRPC compiler plugin.
Building the gRPC compiler plugin is not necessary for Linux, MacOS and
Windows platform as Google provides pre-build versions for them.
For other platforms, such as FreeBSD, building the plugin is a necessary
step. It is also a step that requires documenting as it doesn't built
out-of-the-box for platforms other than Linux, MacOS and Windows.

In addition, the `protoc` compiler needs to be installed into to local
Maven repository. For some platforms, such as FreeBSD, this is an
explicit step. That step has been documented too.