Generate wsdl2java generated sources in the build dir.
This is the default and works better by not having to define custom
cleaning options, carefully defining what to clean and what not. Reason
for deviating from it in the first place was lost familiarity with
Java/Maven ecosystem. Slowly regaining it.

Anyway, now both gRPC/protobuf and CXF generate their artifacts in the
build directory (`target/*`).