Split Service Definition msg from Provider msgs.
NSI is an extremely flexible protocol that is not fully defined using
its WSDL spec. Instead it relies on additional XML schema definitions,
such as those for the Service Definitions, that it can dynamically refer
to. In practise, only one Service Definition is in use (that of
Point2Point connection). Another additional XML schema definition it
relies on is that of the `pathTrace` element that plays a role during
reservation phase of the NSI protocol.

To emphasize the distinction between what is core NSI and what comes
from additional referenced XML schema definitions, we will store the
corresponding protobuf definitions for those additional XML schema
definitions in separate protobuf files. This commit has done that for
the Point2Point Service Definition.

Additionally renamed the `ero` field to `eros` to match protobuf style
guide (use pluralized names for repeated fields).