Change Java package names of protobuf definitions.
We are developing the protobuf version of the NSI protocol as an
hopefully more current and wider applicable option (broader
implementation language support) compared to the SOAP based original. As
such it does not make sense to predefine Java package names to be NREN
and application specific. Hence instead of using `nl.surf.polynsi`
(=NREN domain + app name) as the prefix we will now be using
`org.ogf.nsi` as the prefix.

At some point in the future we should probably extract the protobuf
definitions into a separate repository independent of the PolyNSI
application. But not before we have demonstrated its viability here
in PolyNSI.