Map policy and service XML schemas to Java classes.
These two additional schemas are currently used in production
deployments of the NSI protocol. To be able to fully process a NSI
`reserve` request we cannot exclusivly rely on the Java code generated
by the `wsdl2java` compiler. As it, is documents/XML snippets that adher
to these two schema's are embedded in the `reserve` request. To process
these XML snippets we need the generated Java classes these schemas are
mapped to.