Parse `PathTrace` header extension.
`PathTrace` is a policy extension that can be included in the SOAP header
of the NSI reservation requests and responses. The `PathTrace` element is
contained within the `anyAttribute` of the SOAP header. As such it is
not automatically unmarshalled by the generated Java code from
`wsdl2java`; after all, anything can reside in the `anyAttribute`. Hence
we need to add the JAXB generated mapping for PathTrace to the
JAXBContext manually and unmarshall explicity based on the element name
(`pathTrace`) we encounter.