Autoconfigure gRPC server and client using `grpc-spring-boot-starter`
Hooking up the generated gRPC server and client into Spring Boot
involves a fair amount of Spring Boot machinery. Fortunately the
`grpc-spring-boot-starter` project has automated that all for us.

Using the most recent release of `grpc-spring-boot-starter` required
bumping the versions of `protobuf` and `grpc` packages and that, in
turn, required building some things from source. As usual we need some
additional work to make that all work under FreeBSD. That work, most
notably how to specify the location of the `protoc` compiler in the
presence of an aggregate pom, has been documented in the `README.md`

With the gRPC client and server easily accessible extended the `reserve`
operation to actually call it's gRPC counterpart and extract the
`connection_id`.