Add protobuf def for `ReserveFailed` service.
In protobuf/gRPC, messages can't share names with services. In SOAP
however, it is possible for a message/type to share a name with an
service/operation. This means we can't directly map SOAP operations to
gRPC services and SOAP types to gRPC messages, at least from a naming
point of view. The pattern we have chosen in gRPC to deal with that, is
that of:

   rpc Whatever(WhateverRequest) returns (WhateverResponse).

So the service retains the name from the SOAP operation (eg `Whatever`).
That what is passed along as an argument to the service will be postfxed
with `Request`, where the SOAP version would not have used a postfix at
all. And that what is returned by the service will be postfixed with
`Response`, where the SOAP version would have used either `Response` or
`Ack`.

In addition to creating the `ReserverFailed` service and corresponding
messages, a new `common` protobuf package was created where we moved the
existing `Header` message to (from the `provider` package) as that's now
shared between the `requester` and `provider` package.

Other newly created and shared messages/enums are:

- `TypeValuePair`
- `ServiceException`
- `ReservationState`
- `ProvisionState`
- `LifecycleState`
- `DataPlaneStatus`
- `ConnectionStates`