Rename ApplicationConfig -> ServerConfig and simplify.
Turns out there's no need for for the explicit servlet dispatcher
registration. One does need the beans for the Connection Provider and
Requester though. If they are not there as beans and instead
instantiated explicitely in the endpoint definitions the `GrpcClient`
registration fails. Still don't fully understand why.

With the simplification made it was easy to defer some previously hard
coded settings from code to `application.properties`