Prevent OffsetDatatime from being wrapped in JAXBElement.
Not sure why it was being wrapped in the first place. However
adding `generateElementProperty="false"` to the `globalBindings` in
`jaxb_bindings.xml` it what it finally solved it.

The rewrite of `DataTypeAdapter` was incidental, but as I like this
version better than the original I've kept it.

Removed the @Generated annotation from
`ConnectionServiceProviderPortImpl`. Should have been done on the first
actualy implementation of one of the class's generated methods, in our
case `reserve`. Generally speaking we remove the `@Generated`
annotations from a class or methods once a method has been implemented.