Add ClientConfig for SOAP client requests.
We were already able to receive SOAP requests. Now we need to make SOAP
calls. For that we need to configure a SOAP client. For now only, theres
is only client for ConnectionRequester.

The URL for the `ConnectionRequester` is configured in the
`application.properties` file and points to PolyNSI itself for debugging
purposes. In an actual deployment this should be pointed to an
aggregator or requester agent.