Move soap->protobuf conversion to Converter class.
In addition, implemented XML extraction of `SessionSecurityAttr` as a
string when generating the Protobuf Header. The `SessionSecurityAttr`
should remain as XML as it contains SAML. And that should eventually be
processed in its native format even on PolyNSI gRPC side (eg: in SuPA).

Documented the new and existing `Converter` static methods using
Javadoc.