Implement `reserveConfirmed` (gRPC->SOAP).
We alread had a `toProtobuf(header)` conversion method; initially needed
for `reserve. For `reserveConfirmed` we had to implement the conversion
method `toSoap(header)`. The previously ignored header attribtute
`sessionSecurityAttribute` now had to be addressed in both `toSoap` and
`toProtobuf`.

`sessionSecurityAttribute` contains SAML; an XML based protocol. There
currently is no need to convert protocols other that NSI to Protobuf, not
even when they are embedded in parts of NSI. Hence the decision was made
to pass the SAML XML from SOAP to Protobuf and back to SOAP verbatim.
The problem with that decision is that Apache CXF converts everything
it receives as SOAP documents to mapped Java objects and vice versa.
There is no easy way to tell Apache CXF to not do this de/serialization
for certain attributes. Hence we had to:

* upon receiving a fully deserialized SOAP document as a Java object,
  serialize the `sessionSecurityAttribute` again as an XML string.
* upon receiving a Protobuf message with a `session_security_attributes`
  as an XML string, deserialize it into an appropriate Java object first,
  before Apache CXF can serialize it again to a string as part of
  serializing the entire request.

This means that were it concerns the
`sessionSecurityAttribute`/`session_security_attributes`
de/serialization happens twice. This is wasteful, but a better solution
would have required even more time to investigate and implement (Apache
CXF interceptors? Complex JAXB binding rules that override those
generated by `wsdl2java`?) Fortunately PolyNSI will be a very low
traffic service where this slight inefficiently will not matter.